// web/api-gen/models/Notebook.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { NotebookCell } from './NotebookCell';
export type Notebook = {
    notebookUuid: string;
    notebookName: string;
    updatedAt: string;
    cells: Array<NotebookCell>;
};